config: use "long int" as required by my version of libconfig
Fixes:

config.c: In function ‘load_config’:
config.c:222: warning: passing argument 3 of ‘config_lookup_int’ from incompatible pointer type
/usr/include/libconfig.h:235: note: expected ‘long int *’ but argument is of type ‘int *’
config.c:245: warning: passing argument 3 of ‘config_lookup_int’ from incompatible pointer type
/usr/include/libconfig.h:235: note: expected ‘long int *’ but argument is of type ‘int *’

+++ wmediumd/config.c
@@ -179,7 +179,7 @@ int load_config(const char *file)
 	config_t cfg, *cf;
 	const config_setting_t *ids, *prob_list, *mat_array, *jammer_s;
 	int count_ids, rates_prob, i, j;
-	int count_value, rates_value;
+	long int count_value, rates_value;
 
 	/*initialize the config file*/
 	cf = &cfg;
